Fixing creating new object bug.

+++ lib/DivanteLtd/PimcoreElasticsearchPlugin/Indexer/Service/IndexerService.php
@@ -64,6 +64,10 @@ class IndexerService
                 continue;
             }
 
+            if (!$this->elasticSearchService->exists($indexer->getIndexName(), $indexer->getType(), $element)) {
+                continue;
+            }
+
             $this->elasticSearchService->delete($indexer->getIndexName(), $indexer->getType(), $element);
         }
     }
